refactor: 사용자 입력값 검증 로직 InputViewValidation 객체로 분리 및 테스트 케이스 분리
