+++ src/main/kotlin/view/InputView.kt
@@ -21,6 +21,7 @@ import lotto.constants.Constants.LOTTO_SIZE
 import lotto.constants.Constants.MAX_NUM
 import lotto.constants.Constants.MIN_NUM
 import lotto.constants.Constants.WINNING_NUM_SEPARATOR
+import lotto.utils.InputViewValidation
 
 class InputView {
 
@@ -44,8 +45,7 @@ class InputView {
 
     private fun getUserAmount(): Int = try {
         val userInput = Console.readLine().trim()
-        validateUserAmount(userInput)
-
+        InputViewValidation.validateUserAmount(userInput)
         println()
 
         userInput.toInt()
@@ -53,74 +53,37 @@ class InputView {
         e.message?.let {
             println(getErrorMessage(it))
         } ?: println()
+
         getUserAmount()
-    }
-
-    fun validateUserAmount(userInput: String) {
-        try {
-            val amount = userInput.toInt()
-
-            if (amount <= 0) throw IllegalArgumentException(EXCEPTION_PURCHASE_NO_MORE_THAN_ZERO)
-            if (amount % LOTTO_PRICE != 0) throw IllegalArgumentException(EXCEPTION_PURCHASE_DIVISION)
-        } catch (e: NumberFormatException) {
-            throw IllegalArgumentException(EXCEPTION_PURCHASE_STRING)
-        }
     }
 
     fun getWinningNumbers(): List<Int> = try {
         val numbers = Console.readLine().trim()
         println()
+
         getWinningNumbersList(numbers)
     } catch (e: IllegalArgumentException) {
         println(getErrorMessage(e.message))
+
         getWinningNumbers()
     }
 
     fun getWinningNumbersList(winningNumbers: String): List<Int> {
         val winningNumbersList = winningNumbers.split(WINNING_NUM_SEPARATOR)
 
-        return getValidatedNumbersList(winningNumbersList)
-    }
-
-    private fun getValidatedNumbersList(winningNumbers: List<String>): List<Int> = try {
-        val winningNumbersList = winningNumbers.map {
-            it.toInt()
-        }
-        validateWinningNumbersList(winningNumbersList)
-
-        winningNumbersList
-    } catch (e: NumberFormatException) {
-        throw IllegalArgumentException(EXCEPTION_WINNING_NUMBERS_TYPE)
-    }
-
-
-    private fun validateWinningNumbersList(winningNumbers: List<Int>) {
-        if (winningNumbers.size != LOTTO_SIZE) throw IllegalArgumentException(EXCEPTION_WINNING_NUMBERS_SIZE)
-        if (winningNumbers.size != winningNumbers.distinct().count()) {
-            throw IllegalArgumentException(EXCEPTION_WINNING_NUMBERS_DUPLICATED)
-        }
-        winningNumbers.onEach {
-            if (it !in MIN_NUM..MAX_NUM) throw IllegalArgumentException(EXCEPTION_WINNING_NUMBERS_RANGE)
-        }
+        return InputViewValidation.getValidatedNumbersList(winningNumbersList)
     }
 
     fun getBonusNumber(winningNumbers: List<Int>): Int = try {
         val bonus = Console.readLine().trim()
         println()
-        val bonusNum = getValidatedBonusNumber(bonus, winningNumbers)
+        val bonusNum = InputViewValidation.getValidatedBonusNumber(bonus, winningNumbers)
+
         bonusNum
     } catch (e: IllegalArgumentException) {
         println(getErrorMessage(e.message))
-        getBonusNumber(winningNumbers)
-    }
 
-    fun getValidatedBonusNumber(bonus: String, winningNumbers: List<Int>): Int = try {
-        val bonusNum = bonus.toInt()
-        if (bonusNum !in MIN_NUM..MAX_NUM) throw IllegalArgumentException(EXCEPTION_BONUS_NUMBER_RANGE)
-        if (bonusNum in winningNumbers) throw IllegalArgumentException(EXCEPTION_BONUS_NUMBER_DUPLICATED)
-        bonusNum
-    } catch (e: NumberFormatException) {
-        throw IllegalArgumentException(EXCEPTION_BONUS_NUMBER_TYPE)
+        getBonusNumber(winningNumbers)
     }
 
     private fun getStepMessage(message: String): String = message
